moved code from pype_info to openpype_version and fixed few bugs
